Rainbow works, but not automatic
I would probably need to implement AutoKeyRepeat in some way or form for my function to work correctly, because implementing a sleep or a gettimeofday() inside a loop to call draw_lines_rainbow_offset() causes infinite hangs in the window, no matter how long the waiting period between draws is. I might need to loop inside of a fork for the rainbow effect to work as is.