SHIT'S GOING OOOOON, BHRUV!
Sorry, got a little carried away.

Rotations work as intended, in every direction.

This project was such a mess from start to finish, I absolutely love it.

So now seems like good a time as any to clean up my project and split
the functions into more logical pieces. Never ever should I ask myself :
in what file is that function ?

[todo]
[ ] - project_point needs to check if given point is within bounds. If
not, don't project the point
[ ] - Bresenham should check if either point is within bounds. If none
are, don't draw the line. If one of the two is within bounds, project
the out-of-bounds point onto the ordinate/abscissa, and then draw the
point